+++ packages/html-elements/src/primitives/View.tsx
@@ -48,11 +48,11 @@ export interface WebViewStyle {
   /** @platform web */
   backgroundOrigin?: 'border-box' | 'content-box' | 'padding-box';
   /** @platform web */
-  backgroundPosition?: string;
+  backgroundPosition?: NativeViewStyle['backgroundPosition'] | string;
   /** @platform web */
-  backgroundRepeat?: string;
+  backgroundRepeat?: NativeViewStyle['backgroundRepeat'] | string;
   /** @platform web */
-  backgroundSize?: string;
+  backgroundSize?: NativeViewStyle['backgroundSize'] | string;
   /** @platform web */
   boxShadow?: string | readonly BoxShadowValue[];
   /** @platform web */